refactor: move storage fingerprints out of idle pool
