+++ crates/runner/src/executor/tests/storage_tests.rs
@@ -171,7 +171,7 @@ fn filter_same_artifact_version_keeps_url_for_mount_repair() {
         artifacts: vec![guest_art("my-art", "v1", Some("https://s3/v1"))],
         cleanup_paths: vec![],
     };
-    let prev = crate::idle_pool::StorageFingerprints {
+    let prev = crate::storage_fingerprints::StorageFingerprints {
         storages: HashMap::new(),
         artifacts: art_fp("/workspace", "my-art", "v1"),
     };
@@ -191,7 +191,7 @@ fn filter_different_artifact_version_keeps_url() {
         artifacts: vec![guest_art("my-art", "v2", Some("https://s3/v2"))],
         cleanup_paths: vec![],
     };
-    let prev = crate::idle_pool::StorageFingerprints {
+    let prev = crate::storage_fingerprints::StorageFingerprints {
         storages: HashMap::new(),
         artifacts: art_fp("/workspace", "my-art", "v1"),
     };
@@ -209,7 +209,7 @@ fn filter_different_artifact_name_keeps_url() {
         artifacts: vec![guest_art("other-art", "v1", Some("https://s3/v1"))],
         cleanup_paths: vec![],
     };
-    let prev = crate::idle_pool::StorageFingerprints {
+    let prev = crate::storage_fingerprints::StorageFingerprints {
         storages: HashMap::new(),
         artifacts: art_fp("/workspace", "my-art", "v1"),
     };
@@ -224,7 +224,7 @@ fn filter_new_artifact_not_in_prev_keeps_url() {
         artifacts: vec![guest_art("my-art", "v1", Some("https://s3/v1"))],
         cleanup_paths: vec![],
     };
-    let prev = crate::idle_pool::StorageFingerprints::default();
+    let prev = crate::storage_fingerprints::StorageFingerprints::default();
     let result = filter_unchanged_storages(&manifest, &prev);
     assert!(result.artifacts[0].archive_url.is_some());
 }
@@ -241,7 +241,7 @@ fn filter_empty_prev_downloads_everything() {
         artifacts: vec![guest_art("my-art", "v1", Some("https://s3/v1"))],
         cleanup_paths: vec![],
     };
-    let prev = crate::idle_pool::StorageFingerprints::default();
+    let prev = crate::storage_fingerprints::StorageFingerprints::default();
     let result = filter_unchanged_storages(&manifest, &prev);
     assert!(result.storages[0].archive_url.is_some());
     assert!(result.artifacts[0].archive_url.is_some());
@@ -261,7 +261,7 @@ fn filter_all_unchanged_nulls_storage_urls_and_keeps_artifact_urls() {
     };
     let mut storages = HashMap::new();
     storages.insert("/data".into(), ("vol-1".into(), "v1".into()));
-    let prev = crate::idle_pool::StorageFingerprints {
+    let prev = crate::storage_fingerprints::StorageFingerprints {
         storages,
         artifacts: art_fp("/workspace", "my-art", "v1"),
     };
@@ -304,7 +304,7 @@ fn filter_two_artifacts_at_different_mount_paths() {
     let mut artifacts = HashMap::new();
     artifacts.insert("/workspace".into(), ("art-a".into(), "v1".into()));
     artifacts.insert("/data".into(), ("art-b".into(), "v1".into()));
-    let prev = crate::idle_pool::StorageFingerprints {
+    let prev = crate::storage_fingerprints::StorageFingerprints {
         storages: HashMap::new(),
         artifacts,
     };
@@ -333,7 +333,7 @@ fn filter_detects_removed_artifacts() {
     let mut artifacts = HashMap::new();
     artifacts.insert("/workspace".into(), ("kept".into(), "v1".into()));
     artifacts.insert("/old".into(), ("removed".into(), "v1".into()));
-    let prev = crate::idle_pool::StorageFingerprints {
+    let prev = crate::storage_fingerprints::StorageFingerprints {
         storages: HashMap::new(),
         artifacts,
     };
@@ -371,7 +371,7 @@ fn filter_computes_cleanup_for_changed_storages() {
         "/home/user/.claude/skills/foo".into(),
         ("skill-foo".into(), "v1".into()),
     );
-    let prev = crate::idle_pool::StorageFingerprints {
+    let prev = crate::storage_fingerprints::StorageFingerprints {
         storages,
         artifacts: HashMap::new(),
     };
@@ -406,7 +406,7 @@ fn filter_detects_removed_storages() {
         "/home/user/.claude/skills/old-skill".into(),
         ("old-skill".into(), "v1".into()),
     );
-    let prev = crate::idle_pool::StorageFingerprints {
+    let prev = crate::storage_fingerprints::StorageFingerprints {
         storages,
         artifacts: HashMap::new(),
     };
@@ -427,7 +427,7 @@ fn filter_changed_artifact_adds_cleanup_path() {
         artifacts: vec![guest_art("my-art", "v2", Some("https://s3/v2"))],
         cleanup_paths: vec![],
     };
-    let prev = crate::idle_pool::StorageFingerprints {
+    let prev = crate::storage_fingerprints::StorageFingerprints {
         storages: HashMap::new(),
         artifacts: art_fp("/workspace", "my-art", "v1"),
     };
@@ -447,7 +447,7 @@ fn filter_changed_artifact_with_null_url_adds_cleanup_path() {
         artifacts: vec![guest_art("my-art", "v2", None)],
         cleanup_paths: vec![],
     };
-    let prev = crate::idle_pool::StorageFingerprints {
+    let prev = crate::storage_fingerprints::StorageFingerprints {
         storages: HashMap::new(),
         artifacts: art_fp("/workspace", "my-art", "v1"),
     };
@@ -469,7 +469,7 @@ fn filter_unchanged_artifact_policy_does_not_force_redownload() {
         )],
         cleanup_paths: vec![],
     };
-    let prev = crate::idle_pool::StorageFingerprints {
+    let prev = crate::storage_fingerprints::StorageFingerprints {
         storages: HashMap::new(),
         artifacts: art_fp("/workspace", "memory", "v1"),
     };
@@ -493,7 +493,7 @@ fn filter_changed_storage_with_null_url_adds_cleanup_path() {
     };
     let mut storages = HashMap::new();
     storages.insert("/data".into(), ("vol-1".into(), "v1".into()));
-    let prev = crate::idle_pool::StorageFingerprints {
+    let prev = crate::storage_fingerprints::StorageFingerprints {
         storages,
         artifacts: HashMap::new(),
     };
@@ -517,7 +517,7 @@ fn filter_unchanged_storage_sets_cached_true() {
     };
     let mut storages = HashMap::new();
     storages.insert("/data".into(), ("vol-1".into(), "v1".into()));
-    let prev = crate::idle_pool::StorageFingerprints {
+    let prev = crate::storage_fingerprints::StorageFingerprints {
         storages,
         artifacts: HashMap::new(),
     };
@@ -546,7 +546,7 @@ fn filter_tainted_paths_force_download_even_when_versions_match() {
         }],
         cleanup_paths: vec![],
     };
-    let prev = crate::idle_pool::StorageFingerprints {
+    let prev = crate::storage_fingerprints::StorageFingerprints {
         storages: HashMap::from([("/workspace/repo".into(), ("repo".into(), "v1".into()))]),
         artifacts: HashMap::from([(
             "/workspace/artifact".into(),
@@ -586,7 +586,7 @@ fn filter_tainted_removed_paths_are_cleaned() {
         artifacts: vec![],
         cleanup_paths: vec![],
     };
-    let prev = crate::idle_pool::StorageFingerprints {
+    let prev = crate::storage_fingerprints::StorageFingerprints {
         storages: HashMap::from([(
             "/workspace/removed-storage".into(),
             ("repo".into(), "v1".into()),
